# server/src/Tgm.Roborally.Server/wwwroot/openapi-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "contact" : {
      "email" : "nbrugger@student.tgm.ac.at",
      "name" : "Nils Brugger"
    },
    "description" : "This api controlls the flow of a game and provides it's data.\nIt is desiged to be RESTfull so the structure works simmilar as file system.\nThe service will run and only work in a local network, `game.host` is the IP of the Computer hosting the game and will be found via a IP scan",
    "license" : {
      "name" : "Apache 2.0"
    },
    "title" : "Robot Rally Game logic engine",
    "version" : "0.1.0"
  },
  "servers" : [ {
    "description" : "Production",
    "url" : "http://game.host/v1"
  }, {
    "description" : "Debug",
    "url" : "http://localhost:5050/v1"
  } ],
...
  "paths" : {
...
    "/games/{game_id}/players/" : {
      "get" : {
        "description" : "Returns the index of all players",
        "operationId" : "getAllPlayers",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "game_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/GameID"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/PlayerID"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error-Message"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "summary" : "Get all players",
        "tags" : [ "Players" ]
      },
      "post" : {
        "description" : "Join the given game.\nYou will get your ID by doing this, if you already in the game you can get your ID again if you lost it.<br>\nThe id is neccessary for any further API calls",
        "operationId" : "join",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "game_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/GameID"
          },
          "style" : "simple"
        }, {
          "description" : "The password of the game if the lobby is password protected",
          "explode" : true,
          "in" : "query",
          "name" : "password",
          "required" : false,
          "schema" : {
            "type" : "string"
          },
          "style" : "form"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/JoinResponse"
                }
              }
            },
            "description" : "Joined"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error-Message"
                }
              }
            },
            "description" : "Wrong/No password"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error-Message"
                }
              }
            },
            "description" : "Not Found"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error-Message"
                }
              }
            },
            "description" : "Not Joinable"
          }
        },
        "summary" : "Join game",
        "tags" : [ "Players" ]
      }
    },
    "/games/{game_id}/players/{player_id}" : {
      "delete" : {
        "description" : "Removes a player from the game. This can be done by the player itsself or by the host.",
        "operationId" : "kickPlayer",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "game_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/GameID"
          },
          "style" : "simple"
        }, {
          "explode" : false,
          "in" : "path",
          "name" : "player_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/PlayerID"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error-Message"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "security" : [ {
          "Player-Access-Token" : [ ]
        }, {
          "Host-token-access" : [ ]
        } ],
        "summary" : "Remove Player",
        "tags" : [ "Players" ]
      },
      "get" : {
        "description" : "Get closer information about the player",
        "operationId" : "getPlayer",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "game_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/GameID"
          },
          "style" : "simple"
        }, {
          "explode" : false,
          "in" : "path",
          "name" : "player_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/PlayerID"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Player"
                }
              }
            },
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error-Message"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "summary" : "Get player",
        "tags" : [ "Players" ]
      },
      "patch" : {
        "description" : "# DEPRECATET\n> This feature is useless in this version. It will be usefull in newer versions\n\nSets the type of robot(s) the player is controlling",
        "operationId" : "chooseRobot",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "game_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/GameID"
          },
          "style" : "simple"
        }, {
          "explode" : false,
          "in" : "path",
          "name" : "player_id",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/PlayerID"
          },
          "style" : "simple"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "items" : {
                  "$ref" : "#/components/schemas/Robots"
                },
                "type" : "array"
              }
            }
          },
          "description" : "The robots assigned to the player"
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error-Message"
                }
              }
            },
            "description" : "Not Found"
          }
        },
        "security" : [ {
          "Player-Access-Token" : [ ]
        } ],
        "summary" : "Set Robots",
        "tags" : [ "Players" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "GameID" : {
        "description" : "**Unique**<br>\nThis is the parameter a game is identified by",
        "example" : 537,
        "format" : "int32",
        "maximum" : 2048,
        "minimum" : 0,
        "title" : "GameID",
        "type" : "integer"
      },
...
      "Error-Message" : {
        "description" : "A message about that went wrong. Usefull to display users a short and usefull prompt",
        "properties" : {
          "message" : {
            "description" : "A short message describing what happened in human words",
            "type" : "string"
          },
          "error" : {
            "description" : "The error/exception",
            "type" : "string"
          }
        },
        "required" : [ "message" ],
        "title" : "Error-Message",
        "type" : "object"
      },
...
      "Player" : {
        "additionalProperties" : false,
        "description" : "A player attending in a game.\n#### Warning\nThis is **not** permanent. It is created and removed with the game (or with you joining and leaving the game)",
        "example" : {
          "on-turn" : false,
          "controlled_entities" : [ null, null ],
          "active" : true,
          "id" : 3
        },
        "properties" : {
          "id" : {
            "description" : "This id uniquely identifys the player (in a game).\n\n **Not** to be confused with the `uid` which is used for authentication",
            "example" : 3,
            "maximum" : 8,
            "minimum" : 0,
            "title" : "PlayerID",
            "type" : "integer"
          },
          "controlled_entities" : {
            "description" : "The list of entities controlled by this player",
            "items" : {
              "$ref" : "#/components/schemas/EntityID"
            },
            "minItems" : 1,
            "type" : "array",
            "uniqueItems" : true
          },
          "on-turn" : {
            "default" : false,
            "description" : "Îf this is true rhe player is able to interact at the moment",
            "type" : "boolean"
          },
          "active" : {
            "default" : true,
            "description" : "Defines if the player is actively playing. If this is false the player does random moves. This is only false if the player disconnects",
            "type" : "boolean"
          }
        },
        "required" : [ "controlled_entities", "id" ],
        "title" : "Player",
        "type" : "object"
      },
      "PlayerID" : {
        "description" : "This id uniquely identifys the player (in a game).\n\n **Not** to be confused with the `uid` which is used for authentication",
        "example" : 3,
        "maximum" : 8,
        "minimum" : 0,
        "title" : "PlayerID",
        "type" : "integer"
      },
      "EntityID" : {
        "description" : "The unique identification of this entity. <br>\n*!!!* This is not the id of the player<br>\nThis value will be autogenerated by the api and is read only",
        "format" : "int32",
        "minimum" : 0,
        "title" : "EntityID",
        "type" : "integer"
      },
...
      "JoinResponse" : {
        "description" : "All the data you need as you joined a game.",
        "example" : {
          "pat" : 8017480,
          "id" : 3
        },
        "properties" : {
          "id" : {
            "description" : "This id uniquely identifys the player (in a game).\n\n **Not** to be confused with the `uid` which is used for authentication",
            "example" : 3,
            "maximum" : 8,
            "minimum" : 0,
            "title" : "PlayerID",
            "type" : "integer"
          },
          "pat" : {
            "description" : "The uid is the key for the joined player. You need this key for authentication",
            "maximum" : 99999999,
            "minimum" : 9999,
            "type" : "integer"
          }
        },
        "required" : [ "id", "pat" ],
        "title" : "JoinResponse",
        "type" : "object"
      },
...
      "Robots" : {
        "enum" : [ "dicer", "tank" ],
        "title" : "Robots",
        "type" : "string"
      },
...
    },
    "securitySchemes" : {
      "Host-token-access" : {
        "description" : "The Host of the game (and of the API) has an unique identificaton granting him game manipulating abilities. By this means it is also possible to have a host and a client at the same machine. \n> This could be used for AI`s (bot players)",
        "in" : "query",
        "name" : "hid",
        "type" : "apiKey"
      },
      "Player-Access-Token" : {
        "description" : "Players need a token to join a lobby/game. If the host sets a password the player needs the password to get the token",
        "in" : "query",
        "name" : "pat",
        "type" : "apiKey"
      }
    }
  }
}